#4941 : All Modules CRUDs implement SoftDelete

+++ app/Http/Controllers/QuestionController.php
@@ -26,7 +26,8 @@ class QuestionController extends Controller
 
         if ($request->ajax()) {
             // $data = InspectionItem::all();
-            $data=QuestionItem::query();     
+            $data=QuestionItem::orderBy('id', 'DESC');    
+
             if ($request->get('scout_report_category_id') != '' && $request->get('scout_report_category_id')!=null) {
                 $data = $data->where('scout_report_category_id',$request->get('scout_report_category_id'));
             }    
@@ -35,7 +36,15 @@ class QuestionController extends Controller
             }   
             if ($request->get('commodity_types') != '' && $request->get('commodity_types')!=null) {
                 $data = $data->whereJsonContains('commodity_types',$request->get('commodity_types'));
-            }    
+            }  
+          
+            if ($request->get('is_deleted_at') != '' && $request->get('is_deleted_at')!=null) {
+              
+                if($request->get('is_deleted_at')=='true'){                   
+                    $data = $data->withTrashed();
+                }
+            }             
+            
             $data = $data->get();
 
             return DataTables::of($data)
@@ -58,17 +67,26 @@ class QuestionController extends Controller
                 </a>';
                     }
                     if ($user_data->hasPermission('questions', 'delete')) {
-                    $delete_button = '<a href="#" data-id="' . route('questions.destroy', $item->id) . '" class="btn btn-icon btn-bg-light btn-active-color-danger btn-sm me-1 delete_item">
-                    <!--begin::Svg Icon | path: icons/duotune/art/art005.svg-->
-                    <span class="svg-icon svg-icon-3">
-						<svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
-							<path d="M5 9C5 8.44772 5.44772 8 6 8H18C18.5523 8 19 8.44772 19 9V18C19 19.6569 17.6569 21 16 21H8C6.34315 21 5 19.6569 5 18V9Z" fill="currentColor"></path>
-							<path opacity="0.5" d="M5 5C5 4.44772 5.44772 4 6 4H18C18.5523 4 19 4.44772 19 5V5C19 5.55228 18.5523 6 18 6H6C5.44772 6 5 5.55228 5 5V5Z" fill="currentColor"></path>
-							<path opacity="0.5" d="M9 4C9 3.44772 9.44772 3 10 3H14C14.5523 3 15 3.44772 15 4V4H9V4Z" fill="currentColor"></path>
-						</svg>
-					</span>
-                    <!--end::Svg Icon-->
-                </a>';
+                    if(is_null($item->deleted_at)){
+                        $delete_button = '<a href="#" data-id="' . route('questions.destroy', $item->id) . '" class="btn btn-icon btn-bg-light btn-active-color-danger btn-sm me-1 delete_item">
+                            <!--begin::Svg Icon | path: icons/duotune/art/art005.svg-->
+                            <span class="svg-icon svg-icon-3">
+                                <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
+                                    <path d="M5 9C5 8.44772 5.44772 8 6 8H18C18.5523 8 19 8.44772 19 9V18C19 19.6569 17.6569 21 16 21H8C6.34315 21 5 19.6569 5 18V9Z" fill="currentColor"></path>
+                                    <path opacity="0.5" d="M5 5C5 4.44772 5.44772 4 6 4H18C18.5523 4 19 4.44772 19 5V5C19 5.55228 18.5523 6 18 6H6C5.44772 6 5 5.55228 5 5V5Z" fill="currentColor"></path>
+                                    <path opacity="0.5" d="M9 4C9 3.44772 9.44772 3 10 3H14C14.5523 3 15 3.44772 15 4V4H9V4Z" fill="currentColor"></path>
+                                </svg>
+                            </span>
+                            <!--end::Svg Icon-->
+                        </a>';
+                    }else{
+                         $delete_button = '<a title="UnDelete" href="#" data-id="' . route('questions.undelete', $item->id) . '" class="btn btn-icon btn-bg-light btn-active-color-danger btn-sm me-1 delete_request">
+                        <span class="svg-icon svg-icon-3"><svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none">
+                        <path d="M14.5 20.7259C14.6 21.2259 14.2 21.826 13.7 21.926C13.2 22.026 12.6 22.0259 12.1 22.0259C9.5 22.0259 6.9 21.0259 5 19.1259C1.4 15.5259 1.09998 9.72592 4.29998 5.82592L5.70001 7.22595C3.30001 10.3259 3.59999 14.8259 6.39999 17.7259C8.19999 19.5259 10.8 20.426 13.4 19.926C13.9 19.826 14.4 20.2259 14.5 20.7259ZM18.4 16.8259L19.8 18.2259C22.9 14.3259 22.7 8.52593 19 4.92593C16.7 2.62593 13.5 1.62594 10.3 2.12594C9.79998 2.22594 9.4 2.72595 9.5 3.22595C9.6 3.72595 10.1 4.12594 10.6 4.02594C13.1 3.62594 15.7 4.42595 17.6 6.22595C20.5 9.22595 20.7 13.7259 18.4 16.8259Z" fill="currentColor"/>
+                        <path opacity="0.3" d="M2 3.62592H7C7.6 3.62592 8 4.02592 8 4.62592V9.62589L2 3.62592ZM16 14.4259V19.4259C16 20.0259 16.4 20.4259 17 20.4259H22L16 14.4259Z" fill="currentColor"/>
+                        </svg></span>
+                        </a>';
+                       }
                     }
                     return $edit_button . " " . $delete_button;
                 })
@@ -306,10 +324,10 @@ class QuestionController extends Controller
      * @param  \App\Models\InspectionItem  $inspectionItem
      * @return \Illuminate\Http\Response
      */
-    public function destroy(QuestionItem $questionItem,$id)
+    public function destroy($id)
     {
         
-        $delete_item = $questionItem->where('id',$id)->delete();
+        $delete_item = QuestionItem::where('id',$id)->delete();
         if ($delete_item) {
             return response()->json([
                 'status' => 1,
@@ -324,4 +342,23 @@ class QuestionController extends Controller
             ]);
         }
     }
+
+    public function undelete($id)
+    {
+        $delete_request=QuestionItem::where('id', $id)->withTrashed()->restore();
+        if ($delete_request) {
+            return response()->json([
+                'status' => 1,
+                'result' => 'Success',
+                'message' => "UnDeleted",
+            ]);
+        } else {
+            return response()->json([
+                'status' => -1,
+                'result' => 'fail',
+                'message' => "Not UnDeleted",
+            ]);
+        }
+    }
+
 }
